+++ libopendmi/src/entity/cache.c
@@ -293,22 +293,22 @@ const dmi_entity_spec_t dmi_cache_spec =
 
 const char *dmi_cache_type_name(dmi_cache_type_t value)
 {
-    return dmi_name_lookup(&dmi_cache_type_names, value);
+    return dmi_name_lookup(&dmi_cache_type_names, (int)value);
 }
 
 const char *dmi_cache_mode_name(dmi_cache_mode_t value)
 {
-    return dmi_name_lookup(&dmi_cache_mode_names, value);
+    return dmi_name_lookup(&dmi_cache_mode_names, (int)value);
 }
 
 const char *dmi_cache_assoc_name(dmi_cache_assoc_t value)
 {
-    return dmi_name_lookup(&dmi_cache_assoc_names, value);
+    return dmi_name_lookup(&dmi_cache_assoc_names, (int)value);
 }
 
 const char *dmi_cache_location_name(dmi_cache_location_t value)
 {
-    return dmi_name_lookup(&dmi_cache_location_names, value);
+    return dmi_name_lookup(&dmi_cache_location_names, (int)value);
 }
 
 dmi_size_t dmi_cache_size(uint16_t value)
