+++ libopendmi/src/entity/cooling-device.c
@@ -134,7 +134,7 @@ const dmi_entity_spec_t dmi_cooling_device_spec =
 
 const char *dmi_cooling_device_type_name(dmi_cooling_device_type_t value)
 {
-    return dmi_name_lookup(&dmi_cooling_device_type_names, value);
+    return dmi_name_lookup(&dmi_cooling_device_type_names, (int)value);
 }
 
 static bool dmi_cooling_device_decode(dmi_entity_t *entity)
