+++ libopendmi/src/entity/dell/parallel-port.c
@@ -219,13 +219,13 @@ const dmi_entity_spec_t dmi_dell_parallel_port_spec =
 const char *dmi_dell_parallel_port_connector_type_name(
         dmi_dell_parallel_port_connector_type_t value)
 {
-    return dmi_name_lookup(&dmi_dell_parallel_port_connector_type_names, value);
+    return dmi_name_lookup(&dmi_dell_parallel_port_connector_type_names, (int)value);
 }
 
 const char *dmi_dell_parallel_port_connector_pinout_name(
         dmi_dell_parallel_port_connector_pinout_t value)
 {
-    return dmi_name_lookup(&dmi_dell_parallel_port_connector_pinout_names, value);
+    return dmi_name_lookup(&dmi_dell_parallel_port_connector_pinout_names, (int)value);
 }
 
 static bool dmi_dell_parallel_port_decode(dmi_entity_t *entity)
