+++ libopendmi/src/entity/memory-array.c
@@ -184,12 +184,12 @@ const dmi_entity_spec_t dmi_memory_array_spec =
 
 const char *dmi_memory_array_location_name(dmi_memory_array_location_t value)
 {
-    return dmi_name_lookup(&dmi_memory_array_location_names, value);
+    return dmi_name_lookup(&dmi_memory_array_location_names, (int)value);
 }
 
 const char *dmi_memory_array_usage_name(dmi_memory_array_usage_t value)
 {
-    return dmi_name_lookup(&dmi_memory_array_usage_names, value);
+    return dmi_name_lookup(&dmi_memory_array_usage_names, (int)value);
 }
 
 static bool dmi_memory_array_decode(dmi_entity_t *entity)
