+++ libopendmi/src/entity/memory-device.c
@@ -648,17 +648,17 @@ const dmi_entity_spec_t dmi_memory_device_spec =
 
 const char *dmi_memory_device_type_name(dmi_memory_device_type_t value)
 {
-    return dmi_name_lookup(&dmi_memory_device_type_names, value);
+    return dmi_name_lookup(&dmi_memory_device_type_names, (int)value);
 }
 
 const char *dmi_memory_device_form_factor_name(dmi_memory_device_form_factor_t value)
 {
-    return dmi_name_lookup(&dmi_memory_device_form_factor_names, value);
+    return dmi_name_lookup(&dmi_memory_device_form_factor_names, (int)value);
 }
 
 const char *dmi_memory_device_tech_name(dmi_memory_device_tech_t value)
 {
-    return dmi_name_lookup(&dmi_memory_device_tech_names, value);
+    return dmi_name_lookup(&dmi_memory_device_tech_names, (int)value);
 }
 
 dmi_size_t dmi_memory_device_size(uint16_t value)
