+++ libopendmi/src/entity/memory-error.c
@@ -128,15 +128,15 @@ const dmi_name_set_t dmi_memory_error_operation_names =
 
 const char *dmi_memory_error_type_name(dmi_memory_error_type_t value)
 {
-    return dmi_name_lookup(&dmi_memory_error_type_names, value);
+    return dmi_name_lookup(&dmi_memory_error_type_names, (int)value);
 }
 
 const char *dmi_memory_error_granularity_name(dmi_memory_error_granularity_t value)
 {
-    return dmi_name_lookup(&dmi_memory_error_granularity_names, value);
+    return dmi_name_lookup(&dmi_memory_error_granularity_names, (int)value);
 }
 
 const char *dmi_memory_error_operation_name(dmi_memory_error_operation_t value)
 {
-    return dmi_name_lookup(&dmi_memory_error_operation_names, value);
+    return dmi_name_lookup(&dmi_memory_error_operation_names, (int)value);
 }
