+++ libopendmi/src/entity/onboard-device.c
@@ -154,7 +154,7 @@ const dmi_entity_spec_t dmi_onboard_device_spec =
 
 const char *dmi_onboard_device_type_name(dmi_onboard_device_type_t value)
 {
-    return dmi_name_lookup(&dmi_onboard_device_type_names, value);
+    return dmi_name_lookup(&dmi_onboard_device_type_names, (int)value);
 }
 
 static bool dmi_onboard_device_decode(dmi_entity_t *entity)
