+++ libopendmi/src/entity/pointing-device.c
@@ -166,12 +166,12 @@ const dmi_entity_spec_t dmi_pointing_device_spec =
 
 const char *dmi_pointing_device_type_name(dmi_pointing_device_type_t value)
 {
-    return dmi_name_lookup(&dmi_pointing_device_type_names, value);
+    return dmi_name_lookup(&dmi_pointing_device_type_names, (int)value);
 }
 
 const char *dmi_pointing_device_iface_name(dmi_pointing_device_iface_t value)
 {
-    return dmi_name_lookup(&dmi_pointing_device_iface_names, value);
+    return dmi_name_lookup(&dmi_pointing_device_iface_names, (int)value);
 }
 
 bool dmi_pointing_device_decode(dmi_entity_t *entity)
