+++ libopendmi/src/entity/slot.c
@@ -774,27 +774,27 @@ const dmi_entity_spec_t dmi_slot_spec =
 
 const char *dmi_slot_type_name(dmi_slot_type_t value)
 {
-    return dmi_name_lookup(&dmi_slot_type_names, value);
+    return dmi_name_lookup(&dmi_slot_type_names, (int)value);
 }
 
 const char *dmi_slot_width_name(dmi_slot_width_t value)
 {
-    return dmi_name_lookup(&dmi_slot_width_names, value);
+    return dmi_name_lookup(&dmi_slot_width_names, (int)value);
 }
 
 const char *dmi_slot_usage_name(dmi_slot_usage_t value)
 {
-    return dmi_name_lookup(&dmi_slot_usage_names, value);
+    return dmi_name_lookup(&dmi_slot_usage_names, (int)value);
 }
 
 const char *dmi_slot_length_name(dmi_slot_length_t value)
 {
-    return dmi_name_lookup(&dmi_slot_length_names, value);
+    return dmi_name_lookup(&dmi_slot_length_names, (int)value);
 }
 
 const char *dmi_slot_height_name(dmi_slot_height_t value)
 {
-    return dmi_name_lookup(&dmi_slot_height_names, value);
+    return dmi_name_lookup(&dmi_slot_height_names, (int)value);
 }
 
 static bool dmi_slot_decode(dmi_entity_t *entity)
